Implement pluggable merger
