+++ packages/fabrix/src/renderers/shared.tsx
@@ -10,7 +10,7 @@ import {
 } from "graphql";
 import { DirectiveAttributes } from "@registry";
 import { FabrixContextType } from "@context";
-import { FieldWithDirective } from "@inferer";
+import { FieldWithDirective } from "@readers/shared";
 import { FabrixComponentData } from "../fetcher";
 
 type FabrixComponentFieldsRendererExtraProps = Partial<DirectiveAttributes> & {
@@ -152,27 +152,34 @@ export const resolveFieldTypesFromTypename = (
   }, {});
 };
 
-export type FieldType =
-  | {
-      type: "Scalar";
-      name: string;
-    }
-  | {
-      type: "Enum";
-      name: string;
-      meta: {
-        values: string[];
-      };
-    }
-  | {
-      type: "Object";
-      name: string;
-    }
-  | {
-      type: "List";
-      innerType: NonNullable<FieldType>;
-    }
-  | null;
+type ScalarType = {
+  type: "Scalar";
+  name: string;
+};
+
+type EnumType = {
+  type: "Enum";
+  name: string;
+  meta: {
+    values: string[];
+  };
+};
+
+type ObjectType = {
+  type: "Object";
+  name: string;
+};
+
+type ListType = {
+  type: "List";
+  innerType: NonNullable<FieldType>;
+};
+
+export type FieldType = ScalarType | EnumType | ObjectType | ListType | null;
+export const defaultFieldType = {
+  type: "Scalar" as const,
+  name: "String",
+};
 
 export const resolveFieldType = (
   field: GraphQLOutputType | GraphQLNullableType,
